some changes, cause states errors
