+++ Content.Server/SS220/ChemicalAdaptation/ChemicalAdaptationSystem.cs
@@ -1,97 +1,9 @@
 // © SS220, An EULA/CLA with a hosting restriction, full text: https://raw.githubusercontent.com/SerbiaStrong-220/space-station-14/master/CLA.txt
 
 using Content.Shared.SS220.ChemicalAdaptation;
-using Robust.Shared.Timing;
 
 namespace Content.Server.SS220.ChemicalAdaptation;
 
 public sealed class ChemicalAdaptation : SharedChemicalAdaptationSystem
 {
-    [Dependency] private readonly IGameTiming _time = default!;
-
-    public override void Update(float frameTime)
-    {
-        base.Update(frameTime);
-
-        var query = EntityQueryEnumerator<ChemicalAdaptationComponent>();
-        while (query.MoveNext(out var ent, out var comp))
-        {
-            List<string> toRemove = [];
-            foreach (var (name, info) in comp.ChemicalAdaptations)
-            {
-                if (info.Duration > _time.CurTime)
-                    continue;
-
-                toRemove.Add(name);
-            }
-
-            foreach (var name in toRemove)
-            {
-                RemoveAdaptation(ent, comp, name);
-            }
-        }
-    }
-
-    public void EnsureChemAdaptation(ChemicalAdaptationComponent comp, string chemId, TimeSpan duration, float modifier, bool refresh)
-    {
-        if (!comp.ChemicalAdaptations.TryGetValue(chemId, out var adapt))
-        {
-            comp.ChemicalAdaptations.Add(chemId, new AdaptationInfo(duration, modifier, refresh));
-            return;
-        }
-
-        adapt.Modifier *= modifier;
-
-        if (refresh)
-            adapt.Duration = _time.CurTime + duration;
-        else
-            adapt.Duration += duration;
-    }
-
-    public void RemoveAdaptation(EntityUid ent, ChemicalAdaptationComponent comp, string chemId)
-    {
-        comp.ChemicalAdaptations.Remove(chemId);
-
-        if (comp.ChemicalAdaptations.Count == 0)
-            RemCompDeferred<ChemicalAdaptationComponent>(ent);
-    }
-
-    public override bool TryModifyValue(EntityUid ent, string reagent, ref int value)
-    {
-        if (!TryComp<ChemicalAdaptationComponent>(ent, out var adaptComp))
-            return false;
-
-        if (!adaptComp.ChemicalAdaptations.TryGetValue(reagent, out var adaptationInfo))
-            return false;
-
-        value = (int)(value * adaptationInfo.Modifier);
-
-        return true;
-    }
-
-    public override bool TryModifyValue(EntityUid ent, string reagent, ref float value)
-    {
-        if (!TryComp<ChemicalAdaptationComponent>(ent, out var adaptComp))
-            return false;
-
-        if (!adaptComp.ChemicalAdaptations.TryGetValue(reagent, out var adaptationInfo))
-            return false;
-
-        value = value * adaptationInfo.Modifier;
-
-        return true;
-    }
-
-    public override bool TryModifyValue(EntityUid ent, string reagent, ref TimeSpan value)
-    {
-        if (!TryComp<ChemicalAdaptationComponent>(ent, out var adaptComp))
-            return false;
-
-        if (!adaptComp.ChemicalAdaptations.TryGetValue(reagent, out var adaptationInfo))
-            return false;
-
-        value = value * adaptationInfo.Modifier;
-
-        return true;
-    }
 }
