+++ Content.Server/SS220/EntityEffects/EntityEffectSystem.SS220.ChemicalAdaptation.cs
@@ -1,6 +1,6 @@
 // © SS220, An EULA/CLA with a hosting restriction, full text: https://raw.githubusercontent.com/SerbiaStrong-220/space-station-14/master/CLA.txt
 
-using Content.Server.SS220.ChemicalAdaptation;
+using Content.Shared.SS220.ChemicalAdaptation;
 using Content.Shared.EntityEffects;
 using Content.Server.SS220.EntityEffects.Effects;
 
@@ -8,7 +8,7 @@ namespace Content.Server.EntityEffects;
 
 public sealed partial class EntityEffectSystem : EntitySystem
 {
-    [Dependency] private readonly ChemicalAdaptation _adaptation = default!;
+    [Dependency] private readonly SharedChemicalAdaptationSystem _adaptation = default!;
     private void OnChemicalAdaptation(ref ExecuteEntityEffectEvent<ChemicalAdaptationEffect> args)
     {
         if (args.Args is not EntityEffectReagentArgs reagentArgs)
@@ -22,6 +22,8 @@ public sealed partial class EntityEffectSystem : EntitySystem
         var refresh = args.Effect.Refresh;
 
         var chem = EnsureComp<ChemicalAdaptationComponent>(reagentArgs.TargetEntity);
-        _adaptation.EnsureChemAdaptation(chem, reagentArgs.Reagent.ID, duration, modifier, refresh);
+        _adaptation.EnsureChemAdaptation((reagentArgs.TargetEntity, chem), reagentArgs.Reagent.ID, duration, modifier, refresh);
+
+        DirtyEntity(reagentArgs.TargetEntity);//not sure about this one
     }
 }
